- ESPD-30 add financial ratio requirements
git-svn-id: https://webgate.ec.europa.eu/CITnet/svn/ESPD/trunk@273 f47dfcff-13ca-4e7d-9efb-eacd05900603

+++ espd-web/src/main/java/eu/europa/ec/grow/espd/business/common/ResponseValueParsers.java
@@ -170,14 +170,14 @@ final class ResponseValueParsers {
         }
     }
 
-    private static class QuantityValueParser implements ResponseValueParser<Integer> {
+    private static class QuantityValueParser implements ResponseValueParser<Double> {
 
         @Override
-        public Integer parseValue(ResponseType responseType) {
+        public Double parseValue(ResponseType responseType) {
             if (responseType.getQuantity() == null || responseType.getQuantity().getValue() == null) {
                 return null;
             }
-            return responseType.getQuantity().getValue().intValue();
+            return responseType.getQuantity().getValue().doubleValue();
         }
     }
 }
